requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
dj-database-url

